Move input_read method to Estimator class
